Add files via upload
Problem Statement - There are many orphanages in our society. Our automation solution is trying to help to connect society and Orphanage home quickly and accurately.
Our Bot, reads information filled by person who wants to give a new life to a kid through CARA,which is a government organization. Bot reads information filled in by the end user and checks if they are eligible to adopt a kid. Asap Bot replies user with an email(with a image containing text) and SMS to phone number intimidating whether they are eligible with a welcoming note.